Update code to work with new data

+++ code/gwas/download_gwas_ss.sh
@@ -24,12 +24,18 @@
 # The maximum running time of the job in hours:mins:sec (converted to 10 minutes):
 #$ -l h_rt=0:20:00
 
+module load CBI 
+#module load miniforge3/24.7.1-0
+#conda activate bmi-206-group
+module load htslib/1.21
+
 # Download relevant gwas summary statistics
+
 
 # Define arrays for URLs, output file names, and disease names
 urls=(
     "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90077001-GCST90078000/GCST90077873/harmonised/34662886-GCST90077873-EFO_0000685.h.tsv.gz"
-    "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90014001-GCST90015000/GCST90014023/harmonised/34012112-GCST90014023-EFO_0001359.h.tsv.gz"
+#    "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90014001-GCST90015000/GCST90014023/harmonised/34012112-GCST90014023-EFO_0001359.h.tsv.gz"
     "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90077001-GCST90078000/GCST90077731/harmonised/34662886-GCST90077731-EFO_1001055.h.tsv.gz"
     "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90077001-GCST90078000/GCST90077821/harmonised/34662886-GCST90077821-EFO_0001065.h.tsv.gz"
     "ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/GCST90077001-GCST90078000/GCST90077793/harmonised/34662886-GCST90077793-HP_0000138.h.tsv.gz"
@@ -45,7 +51,7 @@ output_dir="output/gwas_ss_filt"
 # Add corresponding output files here
 output_files=(
               "$data_dir/ra_uk_bb.h.tsv"
-              "$data_dir/t1d_uk_bb.h.tsv"
+#              "$data_dir/t1d_uk_bb.h.tsv"
               "$data_dir/hypo_uk_bb.h.tsv"
               "$data_dir/endo_uk_bb.h.tsv"
               "$data_dir/ovary_cys_uk_bb.h.tsv"
@@ -58,7 +64,7 @@ output_files=(
  # Add corresponding disease names here                  
 disease_names=(
               "Rheumatoid Arthritis"
-              "Type 1 Diabetes"
+#              "Type 1 Diabetes"
               "Hypothyroidism"
               "Endometriosis"
               "Ovary Cysts"
@@ -75,12 +81,30 @@ for i in "${!urls[@]}"; do
 
     # Download and unzip the data
     echo " "
+    echo "Now processing summary statistics for ${disease_name}"
+    echo "output_file: ${output_file}"
+    echo "url: ${output_file}"
     echo " "
-    echo "Downloading data for ${disease_name}..."
-    wget -nc -O "${output_file}.gz" "$url"
+    echo " "
+
+    if [ ! -f "${output_file}.gz" ]; then
+        wget -nc -O "${output_file}.gz" "$url"
+    
+       if [[ "$output_file" == *"age_meno"* &&  ! -f "${output_file}.bgz" ]]; then
+           echo "Downloading data for ${disease_name}..."
+           wget -nc -O "${output_file}.bgz" "$url"
+       fi
+    fi
     
     if [ ! -f "$output_file" ]; then
+       echo " "
+       echo " "
+       echo "Unzip data for ${disease_name}..."
+       if [[ "$output_file" == *"age_meno"* ]]; then
+       bgzip -d "${output_file}.bgz"
+       else
        gunzip "${output_file}.gz"
+       fi
     fi
     
     filt_file="$(echo "$output_file" | sed "s|^$data_dir|$output_dir|" | sed 's|\.tsv$|.filt.tsv|')"
@@ -96,7 +120,9 @@ for i in "${!urls[@]}"; do
     
     if [[ "$output_file" == *"uk"* ]]; then
         if [[ "$output_file" == *"age_meno"* ]]; then #menopause files have a different file structure
-            awk -F'\t' 'NR == 1 || $1 == 6 || $1 == 4 || $1 == 11' "$output_file" > "$filt_file"
+            awk ' NR == 1 || $1 ~ /^6/ || $1 ~ /^4/  || $1 ~ /^11/ ' "$output_file" > "$filt_file"
+
+            #awk -F'\t' 'NR == 1 || $1 == 6 || $1 == 4 || $1 == 11' "$output_file" > "$filt_file"
         else 
         # Filtering and counting based on chromosome and position for "ukbb" files
             awk -F'\t' 'NR == 1 || $3 == 6 || $3 == 4 || $3 == 11' "$output_file" > "$filt_file"
@@ -104,14 +130,14 @@ for i in "${!urls[@]}"; do
         fi
         
         # Print statistics
-        echo "${disease_name} number of sites tested for chromosome 6"
-        awk -F'\t' 'NR == 1 || $3 == 6' "$filt_file" | wc -l
+        echo "${disease_name} number of sites tested for chromosome 4, 6, 11 "
+        "$filt_file" | wc -l
 
-        echo "${disease_name} number of sites for chromosome 4"
-        awk -F'\t' 'NR == 1 || $3 == 4' "$filt_file" | wc -l
-
-        echo "${disease_name} number of sites for chromosome 11"
-        awk -F'\t' 'NR == 1 || $3 == 11' "$filt_file" | wc -l
+        # echo "${disease_name} number of sites for chromosome 4"
+        # awk -F'\t' 'NR == 1 || $3 == 4' "$filt_file" | wc -l
+        # 
+        # echo "${disease_name} number of sites for chromosome 11"
+        # awk -F'\t' 'NR == 1 || $3 == 11' "$filt_file" | wc -l
         
     elif [[ "$output_file" == *"finngen"* ]]; then
     
@@ -119,14 +145,16 @@ for i in "${!urls[@]}"; do
         awk -F'\t' 'NR == 1 || $1 == 6 || $1 == 4 || $1 == 11' "$output_file" > "$filt_file"
         
         # Print statistics
-        echo "${disease_name} number of sites tested for chromosome 6"
-        awk -F'\t' 'NR == 1 || $1 == 6' "$filt_file" | wc -l
-
-        echo "${disease_name} number of sites for chromosome 4"
-        awk -F'\t' 'NR == 1 || $1 == 4' "$filt_file" | wc -l
-
-        echo "${disease_name} number of sites for chromosome 11"
-        awk -F'\t' 'NR == 1 || $1 == 11' "$filt_file" | wc -l
+        echo "${disease_name} number of sites tested for chromosome 4, 6, 11 "
+        "$filt_file" | wc -l
+        # echo "${disease_name} number of sites tested for chromosome 6"
+        # awk -F'\t' 'NR == 1 || $1 == 6' "$filt_file" | wc -l
+        # 
+        # echo "${disease_name} number of sites for chromosome 4"
+        # awk -F'\t' 'NR == 1 || $1 == 4' "$filt_file" | wc -l
+        # 
+        # echo "${disease_name} number of sites for chromosome 11"
+        # awk -F'\t' 'NR == 1 || $1 == 11' "$filt_file" | wc -l
     fi  
 
     echo "Processing complete for ${disease_name}."
